Move agent functionality from http (#153917)
